Fix labels and lay resistors flat.
Left inputs to the line driver were backwards, because the datasheet had
an error. I think I "fixed"it before to match the datasheet.

I also decided to lay the resistors flat. They don't take up that much
room flat, and they will be easier to make pretty on the board, and more
stable with all the vibration.

Also updated the BOM and netlist.