Add a line in module and use a smaller jack.
Lots of changes:
Added a line in. This requires another audio switch, so that's added to
the audio switch module.
Also, since I'm using two jacks, I chose a smaller jack module. This
needed a new footprint, and a new 3d model. The footprint is trickier,
since it really needed slots, which OSH won't guarantee. I've given it a
unique footprint to try to optimize the pad size around the edges of the
pins.