Reduce the line driver back to simple
After determining the gain needed to be 0.735 (which is out of spec for
the amp), I was unsure what to do with the 2nd order filter on the amp.
It wasn't even present on the headphone amp from before, and I suspect
the head unit in the car does some filtering of its own. To simplify
things, I removed the second order filter and went back to a simple line
driver. I also set the values of the gain setting resistors to 8.2 and
11, which gets as close as I can to 0.735 gain using single E24
resistors.