Separate audio ground from power ground
The connection between the CD player and stereo has a separate line for
audio ground and the chassis ground. I didn't previously understand why
this was necessary, so I tied them together. They were tied together in
the stereo anyway. The reason is that when using power, there must be
current returning on the ground line, and since that ground line is not
free of resistance, that causes a slight voltage change. That voltage
difference causes a disagreement in what ground is, and at audio levels,
that makes a significant difference. By isolating the audio ground to a
separate wire that doesn't carry current, we keep the ground potential
the same between the devices, and the audio comes out much cleaner.

On the main connector, I've kept the chassis ground as the power ground,
and only connected the audio ground to the mid-rail on the line driver,
and the line-in ground, as those are the places where audio ground
potential will matter.

As I now understand the grounding situation better, I'm removing the
ground jumper pads.

I will also consider removing the voltage regulator if I can get enough
power out of the regulator on board with the MCU.