Add LEDs to BT module
These already exist on the breakout board, but they will be more visible
if I add resistors to the front panel. They can be removed if space is
limited and the case is translucent, but add them for now. In V2, there
won't be a breakout board, so adding them now makes the V1 to V2 fewer
changes.