jbebels attempt at routing
All power and audio is routed, as is the crystal and UART lines. Other
digital lines remain, and the resistors for the LEDs are moved off-board
until I figure out the best place for them.

This also required a small fix to the schematic. A junction was missing,
causing the audio grounds to not be connected to the cd/stereo audio
ground.