First full layout
I renumbered pins on the MCU so that the layout was cleaner, hence the
regenerated net file.

This passes all the design rules and nothing is unconnected. I managed
to get the ground pour connected without having to pour between pins,
but it seems even better connected if I do, and it is within the design
rules, so it seems safe.